ai/pydantic-ai/func/pyproject.toml
[project]
name = "func"
version = "0.1.0"
description = "Example of AI using custom function."
readme = "../../README.md"
requires-python = ">=3.13"
dependencies = [
    "pydantic>=2.10.4",
    "pydantic-ai>=0.0.17",
    "python-dotenv>=1.0.1",
]
